Migrate account tests
